requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
beautifulsoup4==4.12.3
requests==2.31.0
pyvis==0.3.2
networkx==3.2.1
pybliometrics==3.5.2
lxml==5.1.0
python-dotenv==1.0.0
plotly==5.18.0
